+++ api/v1/toil/resources/healthcheck.ep.toil.arbeit.inc.php
@@ -3,34 +3,54 @@ namespace Toil {
     require_once $_SERVER["DOCUMENT_ROOT"] . "/api/v1/inc/arbeit.inc.php";
     use Toil\EP;
 
-    class healthcheck implements EPInterface {
-        public function __construct(){
-
+    class healthcheck implements EPInterface
+    {
+        public function __construct()
+        {
+            /**
+             * Empty, required by interface
+             *
+             */
         }
 
-        public function __set($name, $value){
+        public function __set($name, $value)
+        {
             $this->$name = $value;
         }
 
-        public function __get($name){
+        public function __get($name)
+        {
             return $this->$name;
         }
 
-        public function get(){
+        public function get()
+        {
             header('Content-Type: application/json');
-            echo json_encode(array("status" => "alive"));
+            echo json_encode(array('status' => "alive"));
         }
 
-        public function post($post = null){
-
+        public function post($post = null)
+        {
+            /**
+             * Empty, required by interface
+             *
+             */
         }
 
-        public function delete(){
-
+        public function delete()
+        {
+            /**
+             * Empty, required by interface
+             *
+             */
         }
 
-        public function put(){
-            
+        public function put()
+        {
+            /**
+             * Empty, required by interface
+             *
+             */
         }
     }
 }